moved user-related DTOs
