+++ src/main/java/edu/oregonstate/cs467/travelplanner/user/service/UserServiceImpl.java
@@ -3,8 +3,8 @@ package edu.oregonstate.cs467.travelplanner.user.service;
 import edu.oregonstate.cs467.travelplanner.experience.service.ExperienceService;
 import edu.oregonstate.cs467.travelplanner.user.model.User;
 import edu.oregonstate.cs467.travelplanner.user.repository.UserRepository;
-import edu.oregonstate.cs467.travelplanner.web.dto.UserProfileDto;
-import edu.oregonstate.cs467.travelplanner.web.dto.UserRegistrationDto;
+import edu.oregonstate.cs467.travelplanner.user.dto.UserProfileDto;
+import edu.oregonstate.cs467.travelplanner.user.dto.UserRegistrationDto;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
