+++ src/test/java/edu/oregonstate/cs467/travelplanner/user/service/UserServiceTest.java
@@ -4,8 +4,8 @@ import edu.oregonstate.cs467.travelplanner.experience.model.Experience;
 import edu.oregonstate.cs467.travelplanner.experience.service.ExperienceService;
 import edu.oregonstate.cs467.travelplanner.user.model.User;
 import edu.oregonstate.cs467.travelplanner.user.repository.UserRepository;
-import edu.oregonstate.cs467.travelplanner.web.dto.UserProfileDto;
-import edu.oregonstate.cs467.travelplanner.web.dto.UserRegistrationDto;
+import edu.oregonstate.cs467.travelplanner.user.dto.UserProfileDto;
+import edu.oregonstate.cs467.travelplanner.user.dto.UserRegistrationDto;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.ArgumentCaptor;
